+++ pywbem/cim_types.py
@@ -21,47 +21,45 @@
 # Author: Ross Peoples <ross.peoples@gmail.com>
 #
 
+# pylint: disable=line-too-long
 """
 Types to represent CIM typed values, and related conversion functions.
 
-The following table shows how CIM typed values are represented as Python
-objects:
+The following table shows how CIM data types are represented in Python.
+Note that some basic CIM data types are represented with built-in Python
+types.
 
-=========================  ===========================
-CIM type                   Python type
-=========================  ===========================
-boolean                    `bool`
-char16                     unicode string or binary bytes, see (1)
-string                     unicode string or binary bytes, see (1)
-string (EmbeddedInstance)  `CIMInstance`
-string (EmbeddedObject)    `CIMInstance` or `CIMClass`
-datetime                   `CIMDateTime`
-reference                  `CIMInstanceName`
-uint8                      `Uint8`
-uint16                     `Uint16`
-uint32                     `Uint32`
-uint64                     `Uint64`
-sint8                      `Sint8`
-sint16                     `Sint16`
-sint32                     `Sint32`
-sint64                     `Sint64`
-real32                     `Real32`
-real64                     `Real64`
-[] (array)                 `list`
-=========================  ===========================
+========================================  =====================================
+CIM data type                             Python type
+========================================  =====================================
+boolean                                   :class:`py2:bool`
+char16                                    `unicode string`_ or `byte string`_
+string                                    `unicode string`_ or `byte string`_
+string (EmbeddedInstance)                 :class:`~pywbem.CIMInstance`
+string (EmbeddedObject)                   :class:`~pywbem.CIMInstance`
+                                          or :class:`~pywbem.CIMClass`
+datetime                                  :class:`~pywbem.CIMDateTime`
+reference                                 :class:`~pywbem.CIMInstanceName`
+uint8                                     :class:`~pywbem.Uint8`
+uint16                                    :class:`~pywbem.Uint16`
+uint32                                    :class:`~pywbem.Uint32`
+uint64                                    :class:`~pywbem.Uint64`
+sint8                                     :class:`~pywbem.Sint8`
+sint16                                    :class:`~pywbem.Sint16`
+sint32                                    :class:`~pywbem.Sint32`
+sint64                                    :class:`~pywbem.Sint64`
+real32                                    :class:`~pywbem.Real32`
+real64                                    :class:`~pywbem.Real64`
+[] (array)                                :class:`py2:list`
+========================================  =====================================
 
-(1) CIM string and char16 types are represented as follows:
-In Python 2 as `unicode` (preferred) or `str`; in Python 3 as `str` (preferred)
-or `bytes`. The implementation may decode binary bytes types offered at the
-interface to unicode text types in the internal representation, using "utf-8"
-encoding.
-
-Note that constructors of PyWBEM classes that take CIM typed values as input
+Note that constructors of pywbem classes that take CIM typed values as input
 may support Python types in addition to those shown above. For example, the
-`CIMProperty` class represents CIM datetime values internally as a
-`CIMDateTime` object, but its constructor accepts `datetime.timedelta`,
-`datetime.datetime`, `str`, and `unicode` (py2 only) objects in addition to
-`CIMDateTime` objects.
+:class:`~pywbem.CIMProperty` class represents property values of CIM datetime
+type internally as :class:`~pywbem.CIMDateTime` objects, but its constructor
+accepts :class:`py2:datetime.timedelta` objects, :class:`py2:datetime.datetime`
+objects, `unicode string`_, and `byte string`_, in addition to
+:class:`~pywbem.CIMDateTime` objects.
 """
 
 # This module is meant to be safe for 'import *'.
@@ -156,19 +154,16 @@ class _CIMComparisonMixin(object): #pylint: disable=too-few-public-methods
 
 
 class MinutesFromUTC(tzinfo):
-
     """
-    A `datetime.tzinfo` implementation defined using a fixed offset in +/-
-    minutes from UTC.
+    An implementation of :class:`py2:datetime.tzinfo` that uses a fixed offset
+    in +/- minutes from UTC.
     """
 
     def __init__(self, offset): # pylint: disable=super-init-not-called
         """
-        Initialize the `MinutesFromUTC` object from a timezone offset.
+        Parameters:
 
-        :Parameters:
-
-          offset : `int`
+          offset (int):
             Timezone offset in +/- minutes from UTC, where a positive value
             indicates minutes east of UTC, and a negative value indicates
             minutes west of UTC.
@@ -177,21 +172,23 @@ class MinutesFromUTC(tzinfo):
 
     def utcoffset(self, dt): # pylint: disable=unused-argument
         """
-        Implement the `datetime.tzinfo.utcoffset` method by returning
-        the timezone offset as a `datetime.timedelta` object.
+        Implementation of the :func:`py2:datetime.tzinfo.utcoffset` method
+        that returns the timezone offset as a :class:`py2:datetime.timedelta`
+        object.
         """
         return self.__offset
 
     def dst(self, dt): # pylint: disable=unused-argument
         """
-        Implement the `datetime.tzinfo.dst` method by returning
-        a DST value of 0 as a `datetime.timedelta` object.
+        Implementation of the :func:`py2:datetime.tzinfo.dst` method
+        that returns a DST value of 0 as a :class:`py2:datetime.timedelta`
+        object.
         """
         return timedelta(0)
 
 
 class CIMType(object):       # pylint: disable=too-few-public-methods
-    """Base type for numeric and datetime CIM types."""
+    """Base type for all CIM data types defined in this package."""
 
     # Note: __str__() is not needed; the inherited method is used,
     # even though there is a __repr__() method here.
@@ -205,7 +202,7 @@ class CIMType(object):       # pylint: disable=too-few-public-methods
 
 class CIMDateTime(CIMType, _CIMComparisonMixin):
     """
-    A value of CIM type datetime.
+    A value of CIM data type datetime.
 
     The object represents either a timezone-aware point in time, or a time
     interval.
@@ -213,24 +210,20 @@ class CIMDateTime(CIMType, _CIMComparisonMixin):
 
     def __init__(self, dtarg):
         """
-        Initialize the `CIMDateTime` object from different types of input
-        object.
+        Parameters:
 
-        :Parameters:
+          dtarg:
+            The value from which the object is initialized, as one of the
+            following types:
 
-          dtarg
-            The input object, as one of the following types:
-
-            * A Unicode string or UTF-8 encoded byte string will be interpreted
-              as CIM datetime format (see DSP0004) and will result in a point
-              in time or a time interval.
-            * A `datetime.datetime` object must be timezone-aware and will
-              result in a point in time.
-            * A `datetime.timedelta` object will result in a time interval.
-            * Another `CIMDateTime` object will be copied.
-
-        :Raises ValueError:
-        :Raises TypeError:
+            * A `unicode string`_ or `byte string`_ object will be interpreted
+              as CIM datetime format (see `DSP0004`_) and will result in a
+              point in time or a time interval.
+            * A :class:`py2:datetime.datetime` object must be timezone-aware
+              and will result in a point in time.
+            * A :class:`py2:datetime.timedelta` object will result in a time
+              interval.
+            * Another :class:`~pywbem.CIMDateTime` object will be copied.
         """
         from .cim_obj import _ensure_unicode # defer due to cyclic deps.
         self.cimtype = 'datetime'
@@ -306,8 +299,8 @@ class CIMDateTime(CIMType, _CIMComparisonMixin):
     @property
     def datetime(self):
         """
-        The point in time represented by the object, as a `datetime.datetime`
-        object.
+        The point in time represented by the object, as a
+        :class:`py2:datetime.datetime` object.
 
         `None` if the object represents a time interval.
         """
@@ -316,8 +309,8 @@ class CIMDateTime(CIMType, _CIMComparisonMixin):
     @property
     def timedelta(self):
         """
-        The time interval represented by the object, as a `datetime.timedelta`
-        object.
+        The time interval represented by the object, as a
+        :class:`py2:datetime.timedelta` object.
 
         `None` if the object represents a point in time.
         """
@@ -349,8 +342,8 @@ class CIMDateTime(CIMType, _CIMComparisonMixin):
     @classmethod
     def now(cls, tzi=None):
         """
-        Factory method that returns a new `CIMDateTime` object representing
-        the current date and time.
+        Factory method that returns a new :class:`~pywbem.CIMDateTime` object
+        representing the current date and time.
 
         The optional timezone information is used to convert the CIM datetime
         value into the desired timezone. That does not change the point in time
@@ -358,16 +351,16 @@ class CIMDateTime(CIMType, _CIMComparisonMixin):
         ``hhmmss`` components of the CIM datetime value to compensate for
         changes in the timezone offset component.
 
-        :Parameters:
+        Parameters:
 
-          tzi : `datetime.tzinfo`
+          tzi (:class:`py2:datetime.tzinfo` or :class:`~pywbem.MinutesFromUTC`):
             Timezone information. `None` means that the current local timezone
-            is used. The `datetime.tzinfo` object may be a `MinutesFromUTC`
-            object.
+            is used.
 
-        :Returns:
+        Returns:
 
-            A new `CIMDateTime` object representing the current date and time.
+            A new :class:`~pywbem.CIMDateTime` object representing the current
+            date and time.
         """
         if tzi is None:
             tzi = MinutesFromUTC(cls.get_local_utcoffset())
@@ -377,8 +370,8 @@ class CIMDateTime(CIMType, _CIMComparisonMixin):
     def fromtimestamp(cls, ts, tzi=None):
         # pylint: disable=invalid-name
         """
-        Factory method that returns a new `CIMDateTime` object from a POSIX
-        timestamp value and optional timezone information.
+        Factory method that returns a new :class:`~pywbem.CIMDateTime` object
+        from a POSIX timestamp value and optional timezone information.
 
         A POSIX timestamp value is the number of seconds since 1970-01-01
         00:00:00 UTC. Thus, a POSIX timestamp value is unambiguous w.r.t. the
@@ -390,20 +383,19 @@ class CIMDateTime(CIMType, _CIMComparisonMixin):
         ``hhmmss`` components of the CIM datetime value to compensate for
         changes in the timezone offset component.
 
-        :Parameters:
+        Parameters:
 
-          ts : `int`
+          ts (int):
             POSIX timestamp value.
 
-          tzi : `datetime.tzinfo`
+          tzi (:class:`py2:datetime.tzinfo` or :class:`~pywbem.MinutesFromUTC`):
             Timezone information. `None` means that the current local timezone
-            is used. The `datetime.tzinfo` object may be a `MinutesFromUTC`
-            object.
+            is used.
 
-        :Returns:
+        Returns:
 
-            A new `CIMDateTime` object representing the specified point in
-            time.
+            A new :class:`~pywbem.CIMDateTime` object representing the
+            specified point in time.
         """
         if tzi is None:
             tzi = MinutesFromUTC(cls.get_local_utcoffset())
@@ -456,74 +448,72 @@ class CIMDateTime(CIMType, _CIMComparisonMixin):
 # CIM integer types
 
 class CIMInt(CIMType, _Longint):
-    """Base type for integer CIM types."""
+    """Base type for integer CIM data types."""
 
 class Uint8(CIMInt):
-    """A value of CIM type uint8."""
+    """A value of CIM data type uint8."""
     cimtype = 'uint8'
 
 class Sint8(CIMInt):
-    """A value of CIM type sint8."""
+    """A value of CIM data type sint8."""
     cimtype = 'sint8'
 
 class Uint16(CIMInt):
-    """A value of CIM type uint16."""
+    """A value of CIM data type uint16."""
     cimtype = 'uint16'
 
 class Sint16(CIMInt):
-    """A value of CIM type sint16."""
+    """A value of CIM data type sint16."""
     cimtype = 'sint16'
 
 class Uint32(CIMInt):
-    """A value of CIM type uint32."""
+    """A value of CIM data type uint32."""
     cimtype = 'uint32'
 
 class Sint32(CIMInt):
-    """A value of CIM type sint32."""
+    """A value of CIM data type sint32."""
     cimtype = 'sint32'
 
 class Uint64(CIMInt):
-    """A value of CIM type uint64."""
+    """A value of CIM data type uint64."""
     cimtype = 'uint64'
 
 class Sint64(CIMInt):
-    """A value of CIM type sint64."""
+    """A value of CIM data type sint64."""
     cimtype = 'sint64'
 
 # CIM float types
 
 class CIMFloat(CIMType, float):
-    """Base type for real (floating point) CIM types."""
+    """Base type for real (floating point) CIM data types."""
 
 class Real32(CIMFloat):
-    """A value of CIM type real32."""
+    """A value of CIM data type real32."""
     cimtype = 'real32'
 
 class Real64(CIMFloat):
-    """A value of CIM type real64."""
+    """A value of CIM data type real64."""
     cimtype = 'real64'
 
 def cimtype(obj):
     """
-    Return the CIM type name of a value, as a string.
+    Return the CIM data type name of a CIM typed object, as a string.
 
-    For an array, the type is determined from the first array element because
-    CIM arrays must be homogeneous. If the array is empty, ValueError is
-    raised.
+    For an array, the type is determined from the first array element
+    (CIM arrays must be homogeneous w.r.t. the type of their elements).
 
-    If the type of the value is not a CIM type, TypeError is raised.
+    Parameters:
 
-    :Parameters:
+      obj (`CIM data type`_):
+        The object whose CIM data type name is returned.
 
-      obj : CIM typed value
-        The value whose CIM type name is returned.
+    Returns:
 
-    :Returns:
+        The CIM data type name of the object, as a string (e.g. ``"uint8"``).
 
-        The CIM type name of the value, as a string.
-
-    :Raises TypeError: Type is not a CIM type.
-    :Raises ValueError: Cannot determine CIM type from empty array.
+    Raises:
+        TypeError: Type is not a CIM data type.
+        ValueError: Cannot determine CIM data type from an empty array.
     """
     if isinstance(obj, CIMType):
         return obj.cimtype
@@ -534,11 +524,11 @@ def cimtype(obj):
         return 'string'
     if isinstance(obj, list):
         if len(obj) == 0:
-            raise ValueError("Cannot determine CIM type from empty array")
+            raise ValueError("Cannot determine CIM data type from an empty array")
         return cimtype(obj[0])
     if isinstance(obj, (datetime, timedelta)):
         return 'datetime'
-    raise TypeError("Type %s of this value is not a CIM type: %r" % \
+    raise TypeError("Type %s of this object is not a CIM data type: %r" % \
                     (type(obj), obj))
 
 _TYPE_FROM_NAME = {
@@ -561,25 +551,28 @@ _TYPE_FROM_NAME = {
 
 def type_from_name(type_name):
     """
-    Return the Python type object for a given CIM type name.
+    Return the Python type object for a given CIM data type name.
 
-    For example, type name `'uint8'` will return type `Uint8`.
+    For example, type name "uint8" will return type object
+    :class:`~pywbem.Uint8`.
 
-    For CIM types `string` and `char16`, the preferred Python type
-    for unicode text representation is returned.
+    For CIM data type names ``"string"`` and ``"char16"``, the
+    `unicode string`_ type is returned (Unicode strings are the preferred
+    representation for these CIM data types).
 
-    :Parameters:
+    Parameters:
 
       type_name : string
-        The simple (=non-array) CIM type name (e.g. `'uint8'` or
-        `'reference'`).
+        The simple (=non-array) CIM data type name (e.g. ``"uint8"`` or
+        ``"reference"``).
 
-    :Returns:
+    Returns:
 
-        The Python type object for the CIM type (e.g. `Uint8` or
-        `CIMInstanceName`).
+        The Python type object for the CIM data type (e.g.
+        :class:`~pywbem.Uint8` or :class:`~pywbem.CIMInstanceName`).
 
-    :Raises ValueError: Unknown CIM type name.
+    Raises:
+        ValueError: Unknown CIM data type name.
     """
     if type_name == 'reference':
         # move import to run time to avoid circular imports
@@ -588,27 +581,28 @@ def type_from_name(type_name):
     try:
         type_obj = _TYPE_FROM_NAME[type_name]
     except KeyError:
-        raise ValueError("Unknown CIM type name: %r" % type_name)
+        raise ValueError("Unknown CIM data type name: %r" % type_name)
     return type_obj
 
 def atomic_to_cim_xml(obj):
     """
-    Convert a value of an atomic scalar CIM type to a CIM-XML Unicode string
-    and return that string.
+    Convert a value of an atomic scalar CIM data type to a CIM-XML string and
+    return that string.
 
     TODO: Verify whether we can change this function to raise a ValueError in
     case the value is not CIM typed.
 
-    :Parameters:
+    Parameters:
 
-      obj : CIM typed value.
-        The CIM typed value`, including `None`. Must be a scalar. Must be an
-        atomic type (i.e. not `CIMInstance` or `CIMClass`).
+      obj (atomic scalar CIM typed value):
+        The CIM typed value, including `None`. Must be a scalar (not an array).
+        Must be an atomic type (i.e. those listed in `CIM data types`_, except
+        any `CIM objects`_).
 
-    :Returns:
+    Returns:
 
-        A Unicode string in CIM-XML value format representing the CIM typed
-        value. For a value of `None`, `None` is returned.
+        A `unicode string`_ object in CIM-XML value format representing the CIM
+        typed value. For a value of `None`, `None` is returned.
     """
     # pylint: disable=too-many-return-statements
     from .cim_obj import _ensure_unicode, _convert_unicode  # due to cycles